+++ corpus/aliased_namespaces/issue_2614.cljs
@@ -1,0 +1,4 @@
+(ns issue-2614
+  (:require [clojure.string :as str]))
+
+(clojure.string/trim "repro")