+++ src/data_readers.clj
@@ -1,1 +1,2 @@
-{f/q clojure.set/join}
+{f/q clojure.set/join
+ dbg clojure.core/identity}